Migrate unit tests of startup-modals service
